Add files via upload
This is my first project. Not completely my idea. Watched a youtube video and copied the code. Then recreated the entire project. Then deleted the whole project and started building it from scratch again. Understood the HTML and some parts of the CSS styling. Then deleted it again. Recreated it again and this time googled everything and understood the entire CSS and JS too. Then deleted it and again recreated it again from scratch without any help and added my personal touch. I don't want to be judged because this is my first project. More projects on the way